refactor: update spar endpoints and conventions
